engine: extract esm/tour.js
